refactor(item): change from package-by-feature to package-by-layer
