--- backend/item-service/src/main/java/com/mazad/item/features/item/service/ItemService.java
@@ -1,16 +1,0 @@
-package com.mazad.item.features.item.service;
-
-import com.mazad.item.features.item.dto.ItemRequest;
-import com.mazad.item.features.item.dto.ItemResponse;
-import org.springframework.data.web.PagedModel;
-
-import java.util.UUID;
-
-public interface ItemService {
-
-    public ItemResponse createItem(ItemRequest itemRequest, UUID sellerId);
-
-    public ItemResponse getItem(Long id);
-
-    public PagedModel<ItemResponse> getItemsPage(int page, int size);
-}
